+++ src/app/page.tsx
@@ -1,10 +1,31 @@
-export default function Home() {
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+export default function HomePage() {
+  const router = useRouter();
+
   return (
-    <div className="bg-prima flex justify-center text-2xl">
-      Butter-Salt-Flour
-      <div className="text-banana">banana</div>
-      <div className="bg bg-wonbinBrown">wonbinBrown</div>
-      <div className="text-banana">banana</div>
-    </div>
+    <main className="flex flex-col items-center justify-center h-screen bg-gray-50 px-4">
+      <section className="text-center space-y-4">
+        <h1 className="text-4xl md:text-5xl font-bold tracking-tight">
+          할매야 놀자
+        </h1>
+        <p className="text-gray-600 text-base md:text-lg max-w-md">
+          고립된 할머니와, 연결이 필요한 청년을 위한 따뜻한 하루 챌린지
+        </p>
+
+        <button
+          onClick={() => router.push('/login')}
+          className="mt-4 px-6 py-2 rounded bg-black text-white hover:bg-gray-800 transition"
+        >
+          로그인하기
+        </button>
+      </section>
+
+      <footer className="absolute bottom-6 text-2xl text-gray-400">
+        “오늘, 할매랑 산책 어때요?”
+      </footer>
+    </main>
   );
 }
